update script and test

+++ build.sh
@@ -1,12 +1,13 @@
 PYBIND11_DIR=$(python3 -m pybind11 --cmakedir)
 export pybind11_DIR="$PYBIND11_DIR" # accessible by child process
 
-rm -r build
-mkdir build
-cd build
-cmake .. 
-cmake --build .
-cd ..
-rm app/key_value_store_module*.so
-cp build/key_value_store_module*.so app
+[ -d build ] && rm -r build
+[ -f app/key_value_store_module*.so ] && rm app/key_value_store_module*.so
+
+mkdir build && \
+cd build && \
+make .. && \
+cmake --build . && \
+cd .. && \
+cp build/key_value_store_module*.so app && \
 fastapi dev app/main.py
